# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenWiki V2 v1 API
  version: 1.0.0
  summary: 企业内部 LLM Wiki 知识库系统 v1 API 契约
  license:
    name: Internal
    identifier: UNLICENSED
  description: |
    本契约覆盖 OpenWiki V2 v1 的 M0-M6 API。

    文档优先级：ROADMAP > PRD > TRD > Architecture。ROADMAP 中的 v1 冻结约束覆盖 TRD 中与 v1 范围不一致的完整形态描述。
servers:
  - url: /api/v1
    description: v1 API prefix
security:
  - bearerAuth: []
...
paths:
...
  /kbs:
    get:
      tags: [Knowledge Bases]
      summary: 知识库列表
      operationId: listKnowledgeBases
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/KnowledgeBaseType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/KnowledgeBaseStatus"
        - name: q
          in: query
          schema:
            type: string
            maxLength: 128
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 知识库分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBasePage"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Knowledge Bases]
      summary: 创建 Source KB 或 Wiki KB
      operationId: createKnowledgeBase
      x-rbac: [admin]
      description: v1 仅允许 Ollama 1024 维 embedding 模型；创建后模型 tag、digest、维度不可通过更新接口修改。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/SourceKnowledgeBaseCreateRequest"
                - $ref: "#/components/schemas/WikiKnowledgeBaseCreateRequest"
      responses:
        "201":
          description: 知识库已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /kbs/{kb_id}:
    get:
      tags: [Knowledge Bases]
      summary: 知识库详情
      operationId: getKnowledgeBase
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          description: 知识库详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Knowledge Bases]
      summary: 更新知识库可变设置
      operationId: updateKnowledgeBase
      x-rbac: [admin]
      description: 不允许修改 embedding_provider、embedding_model_tag、embedding_model_digest、embedding_dim。
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseUpdateRequest"
      responses:
        "200":
          description: 更新后的知识库
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Knowledge Bases]
      summary: 删除知识库
      operationId: deleteKnowledgeBase
      x-rbac: [admin]
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "204":
          description: 删除成功
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /kbs/{kb_id}/bindings:
    post:
      tags: [Knowledge Bases]
      summary: 为 Wiki KB 绑定 Source KB
      operationId: bindSourceKnowledgeBase
      x-rbac: [admin]
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WikiSourceBindingRequest"
      responses:
        "201":
          description: 绑定成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiSourceBinding"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /kbs/{kb_id}/bindings/{source_kb_id}:
    delete:
      tags: [Knowledge Bases]
      summary: 解绑 Source KB
      operationId: unbindSourceKnowledgeBase
      x-rbac: [admin]
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: source_kb_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "204":
          description: 解绑成功
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /kbs/{kb_id}/chunk-preview:
    post:
      tags: [Knowledge Bases]
      summary: 分块预览
      operationId: previewChunks
      x-rbac: [admin]
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkPreviewRequest"
      responses:
        "200":
          description: 分块预览结果
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChunkPreviewItem"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /kbs/{kb_id}/documents/upload:
    post:
      tags: [Documents]
      summary: 上传 Markdown/TXT 文档
      operationId: uploadDocuments
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                tag_ids:
                  type: array
                  items:
                    $ref: "#/components/schemas/Uuid"
      responses:
        "202":
          description: 已创建文档和处理任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 同一 KB 内存在相同 SHA-256 文件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /kbs/{kb_id}/documents:
    get:
      tags: [Documents]
      summary: 文档列表
      operationId: listDocuments
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: q
          in: query
          schema:
            type: string
            maxLength: 128
        - name: tag_id
          in: query
          schema:
            $ref: "#/components/schemas/Uuid"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DocumentStatus"
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, filename_asc, filename_desc]
            default: created_at_desc
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 文档分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPage"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{document_id}:
    get:
      tags: [Documents]
      summary: 文档详情
      operationId: getDocument
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: 文档详情，包含原文预览和 chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Documents]
      summary: 删除文档
      operationId: deleteDocument
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: 删除成功，并清理 chunks 与向量索引
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{document_id}/retry:
    post:
      tags: [Documents]
      summary: 重试失败文档处理
      operationId: retryDocument
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "202":
          description: 已创建新的重试任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAcceptedResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /kbs/{kb_id}/tags:
    get:
      tags: [Tags]
      summary: 标签列表
      operationId: listTags
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          description: 标签列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Tags]
      summary: 创建标签
      operationId: createTag
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateRequest"
      responses:
        "201":
          description: 标签已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /kbs/{kb_id}/tags/{tag_id}:
    patch:
      tags: [Tags]
      summary: 重命名标签
      operationId: updateTag
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/TagId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateRequest"
      responses:
        "200":
          description: 更新后的标签
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Tags]
      summary: 删除标签
      operationId: deleteTag
      x-rbac: [admin, editor]
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: 删除成功
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    TagId:
      name: tag_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
...
    Forbidden:
      description: 无权限或非团队成员
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 资源状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: 请求参数或请求体校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
...
    KnowledgeBaseType:
      type: string
      enum: [document, wiki]
    KnowledgeBaseStatus:
      type: string
      enum: [active, building, disabled, embedding_incompatible]
    DocumentStatus:
      type: string
      enum: [pending, running, completed, failed]
...
    ChunkType:
      type: string
      enum: [text, wiki_page]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              examples: [validation_error]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
    PageMeta:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
...
    ChunkingConfig:
      type: object
      required: [chunk_size, chunk_overlap, strategy]
      properties:
        chunk_size:
          type: integer
          minimum: 128
          maximum: 4096
          default: 512
        chunk_overlap:
          type: integer
          minimum: 0
          maximum: 1024
          default: 80
        strategy:
          type: string
          enum: [header_aware]
          default: header_aware
    WikiConfig:
      type: object
      required: [auto_ingest]
      properties:
        auto_ingest:
          type: boolean
          default: false
        llm_timeout_seconds:
          type: integer
          minimum: 1
          default: 60
        llm_max_retries:
          type: integer
          minimum: 0
          maximum: 5
          default: 3
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
    SourceKnowledgeBaseCreateRequest:
      type: object
      required: [type, name, embedding_model_tag]
      properties:
        type:
          const: document
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        embedding_model_tag:
          type: string
        chunking_config:
          $ref: "#/components/schemas/ChunkingConfig"
    WikiKnowledgeBaseCreateRequest:
      type: object
      required: [type, name, embedding_model_tag, source_knowledge_base_ids]
      properties:
        type:
          const: wiki
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        embedding_model_tag:
          type: string
        source_knowledge_base_ids:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Uuid"
        wiki_config:
          $ref: "#/components/schemas/WikiConfig"
    KnowledgeBaseUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [active, disabled]
        chunking_config:
          $ref: "#/components/schemas/ChunkingConfig"
        wiki_config:
          $ref: "#/components/schemas/WikiConfig"
    KnowledgeBase:
      type: object
      required:
        - id
        - workspace_id
        - name
        - type
        - status
        - embedding_provider
        - embedding_model_tag
        - embedding_model_digest
        - embedding_dim
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        workspace_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/KnowledgeBaseType"
        status:
          $ref: "#/components/schemas/KnowledgeBaseStatus"
        embedding_provider:
          type: string
          const: ollama
        embedding_model_tag:
          type: string
        embedding_model_digest:
          type: string
        embedding_dim:
          type: integer
          const: 1024
        chunking_config:
          $ref: "#/components/schemas/ChunkingConfig"
        wiki_config:
          $ref: "#/components/schemas/WikiConfig"
        document_count:
          type: integer
          minimum: 0
        page_count:
          type: integer
          minimum: 0
        bound_source_kbs:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBaseSummary"
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
    KnowledgeBaseSummary:
      type: object
      required: [id, name, type, status]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        type:
          $ref: "#/components/schemas/KnowledgeBaseType"
        status:
          $ref: "#/components/schemas/KnowledgeBaseStatus"
    KnowledgeBasePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/KnowledgeBase"
    WikiSourceBindingRequest:
      type: object
      required: [source_kb_id]
      properties:
        source_kb_id:
          $ref: "#/components/schemas/Uuid"
    WikiSourceBinding:
      type: object
      required: [id, wiki_kb_id, source_kb_id, created_at]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        wiki_kb_id:
          $ref: "#/components/schemas/Uuid"
        source_kb_id:
          $ref: "#/components/schemas/Uuid"
        created_at:
          $ref: "#/components/schemas/DateTime"
    Tag:
      type: object
      required: [id, kb_id, name]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        kb_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
    TagCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
    Document:
      type: object
      required: [id, kb_id, filename, file_hash, file_size, status, chunk_count, created_by, created_by_username, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        kb_id:
          $ref: "#/components/schemas/Uuid"
        filename:
          type: string
        file_hash:
          type: string
          minLength: 64
          maxLength: 64
        file_size:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/DocumentStatus"
        error_message:
          type:
            - string
            - "null"
        chunk_count:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        created_by:
          $ref: "#/components/schemas/Uuid"
        created_by_username:
          type: string
          minLength: 1
          maxLength: 64
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
    DocumentPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Document"
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          required: [content, chunks]
          properties:
            content:
              type: string
              description: 原始 Markdown/TXT 文本
            chunks:
              type: array
              items:
                $ref: "#/components/schemas/Chunk"
    DocumentUploadResponse:
      type: object
      required: [documents, task_ids]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        task_ids:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
    Chunk:
      type: object
      required: [id, kb_id, content, header_path, seq, start_pos, end_pos, chunk_type, created_at]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        document_id:
          $ref: "#/components/schemas/Uuid"
        kb_id:
          $ref: "#/components/schemas/Uuid"
        content:
          type: string
        header_path:
          type: array
          items:
            type: string
        seq:
          type: integer
          minimum: 0
        start_pos:
          type: integer
          minimum: 0
        end_pos:
          type: integer
          minimum: 0
        chunk_type:
          $ref: "#/components/schemas/ChunkType"
        source_page_id:
          $ref: "#/components/schemas/Uuid"
        created_at:
          $ref: "#/components/schemas/DateTime"
    ChunkPreviewRequest:
      type: object
      required: [content, content_type]
      properties:
        content:
          type: string
          minLength: 1
        content_type:
          type: string
          enum: [markdown, text]
        chunking_config:
          $ref: "#/components/schemas/ChunkingConfig"
    ChunkPreviewItem:
      type: object
      required: [content, header_path, seq, start_pos, end_pos, char_count]
      properties:
        content:
          type: string
        header_path:
          type: array
          items:
            type: string
        seq:
          type: integer
        start_pos:
          type: integer
        end_pos:
          type: integer
        char_count:
          type: integer
    TaskAcceptedResponse:
      type: object
      required: [task_id]
      properties:
        task_id:
          $ref: "#/components/schemas/Uuid"
...